pyproject.toml
[tool.poetry]
name = "gpt-eval"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.24.4"
datasets = "^2.14.5"
easyllm = "^0.5.0"
openai = "^0.28.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
